Added hook for sys_getdents64
It is a copy-paste of the hook of sys_getdents, just changing
the struct linux_dirent by linux_dirent64. I will figure
out how to fix this ugliness. Also, I had to put a byte of
padding inside the linux_dirent64 structure to make it work
correctly, not sure why