Improved hooks definition macros
Now there's a single definition of hook_define, and differences between kernel versions
are defined in DECL and ARGS macros for each function called in hook_define.
In order to implement a simple hook, with hook_define the only thing needed is
implementing the do_hook function and performing the hook in hooks_init.